requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
openai>=1.30.0
faiss-cpu>=1.8.0
numpy>=1.26.0
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.2.1
requests>=2.31.0
httpx>=0.27.0
jinja2>=3.1.3
python-multipart>=0.0.9
